Add a dummy login form
Add a simple login handler displaying a form and handling its data,
currently just responding with the username when the password is the
same.

Add a little template engine, HTML escaping functions, UTF8 sanitizer
code, a middleware parsing form data and a tool to compile in static
HTML templates to support this.