CSRFProtect: Add CSRF protection
Add a new middleware that can create a random token automatically stored
in the session, and transparently intercepts every posted form to verify
it contains the latest token stored in the current session, therefore
enforcing only the user agent that last requested the form can
successfully POST it.

Add and fill hidden token field to both login and logout form.