HttpServer: Translate NAT64 to IPv4 addresses
Add IpAddr class parsing strings to IP addresses or networks (with a
prefix length) and a function to map a NAT64 address back to the plain
IPv4 address using given NAT64 prefixes. Use this in HttpServer to
automatically do the translation using one configurable NAT64 prefix.

Add a server property in configuration file for that, default to the
well-known prefix 64:ff9b::/96.