RateLimit: Add configurable rate limiter
Allow configuration of several time intervals in seconds and a maximum
count for each of them. Internally divide an interval in up to 512
equally sized periods with a maximum resolution of 1 second and maintain
individual counters for each of these periods. Optionally use a mutex to
lock the rate limiter for usage from multiple threads.