Password: Make sure to print prompts to terminal
For password input from the terminal, make sure to print prompts to this
same terminal. For this, first try to print them to standard input(!) ­
which *should* succeed if it is connected to the terminal, which we
already checked earlier.

If this fails nevertheless, try opening our own fd for the terminal.
If this fails as well, fall back to standard error, but only if this is
also connected to a terminal, otherwise bail out.