pow.mjs: Improve performance
Speed up the pseudo base64 encoding: bit shifting and masking seems to
be a lot faster than division and modulo, it brings a million iterations
on my machine in Chromium down from ~220ms to ~80ms.

Apply similar tweaks for obtaining a nibble from the hash candidate, and
remove the check for the hash length as this can't happen (unless
there's a bug in the hashing itself).