Authenticator: Full reconfiguration support
Refactor to allow updating when registering a realm or a default failed
login limit. Add realm and limit versions to the auth info in the
session, so users are automatically logged out for a changed realm and
the rate limiter for failed logins is automatically re-created when
configuration changes.

Add support in Config to learn about removed realms and update the list
of configured realms.