Config: Fix command-line override on reload
For settings that can be given both on the commandline and in the config
file, make sure the commandline always overrides the config file, even
on reload. This affects uid, gid and resolveHosts, but only resolveHosts
needs special care here, because uid and gid cannot be changed at
runtime anyways.