Session: Reduce usage, make explicit
Instead of always creating a session directly from the middleware, only
look for an existing session and make creation explicit with
Session_start().

Use this to only use a session when it is really required. Drop storing
parameters (realm, redirect uri) in the session at all, instead use
query paramters to pass them in the initial redirect to the login form,
and hidden form fields to pass them to the handler processing the POSTed
login form.

Create a session when a user actually logs in, either to create and use
the rate limiter for unsuccessful logins, or to store the user info for
successful logins. Also force creation of a session in the pow checker,
to store the challenge given to the client.

This completely avoids sessions for most "bot traffic".